test: split request handler firewall dispatch tests
